Pingpong example for OSM-to-vim-emu deployments.
This commit cleans up the outdated emulator examples
from the SONATA project and adds example VNFs (ping and pong)
as well as a corresponding NSD as an up-to-date
example that is tested with OSM rel. THREE.

The added example service and VNFs is used by the
end-to-end example documented in the OSM wiki.
Goal is to give users an pre-defined example on how to
on-board and instantiate a network service on the
emulator using OSM.

Also improved README.md.

Change-Id: I45a00f1b6acf1664a96b22e21be4c23612f581cf
Signed-off-by: peusterm <manuel.peuster@uni-paderborn.de>